Develop

+++ asiic-backend/src/main/java/com/assic/muni/domain/model/AsTabla.java
@@ -9,8 +9,6 @@ import jakarta.validation.constraints.Size;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 @Entity
@@ -18,12 +16,11 @@ import java.math.BigDecimal;
 public class AsTabla {
     @Id
     @Column(name = "ta_id", nullable = false, precision = 2)
-    private BigDecimal id;
+    private Short id;
 
     @Size(max = 45)
     @NotNull
     @Column(name = "ta_nombre", nullable = false, length = 45)
     private String taNombre;
-
 
 }
